requirements.txt
python-telegram-bot
requests
websocket-client
